example/udp_loopback: simplify/cleanup and make it more generic
